# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TaskFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      tags: [categories]
      operationId: listCategories
      summary: List all categories for current user
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      tags: [categories]
      operationId: createCategory
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /categories/{id}:
    patch:
      tags: [categories]
      operationId: updateCategory
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [categories]
      operationId: deleteCategory
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
  /dashboard/activity:
    get:
      tags: [dashboard]
      operationId: getDashboardActivity
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Task schemas ────────────────────────────────────────────────────────
...
    Category:
      type: object
      required: [id, userId, name, color, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        color:
          type: string
        icon:
          type: ["string", "null"]
        taskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
          minLength: 1
        icon:
          type: ["string", "null"]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
          minLength: 1
        icon:
          type: ["string", "null"]

    # ── Dashboard schemas ────────────────────────────────────────────────────
...
    ActivityItem:
      type: object
      required: [id, action, taskTitle, timestamp]
      properties:
        id:
          type: integer
        action:
          type: string
          enum: [created, completed, updated, deleted, archived, restored]
        taskTitle:
          type: string
        taskId:
          type: ["integer", "null"]
        timestamp:
          type: string
          format: date-time
...