# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemAction } from './activityItemAction';

export interface ActivityItem {
  id: number;
  action: ActivityItemAction;
  taskTitle: string;
  /** @nullable */
  taskId?: number | null;
  timestamp: Date;
}