# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  color?: string;
  /** @nullable */
  icon?: string | null;
}